move class SimpleNetwork + ResidualNetwork to new module aviary/nn.py
