+++ aviary/core.py
@@ -25,6 +25,7 @@ if sys.version_info < (3, 8):
     from typing_extensions import Literal
 else:
     from typing import Literal
+
 try:
     import wandb
 except ImportError:
